Update and rename FileDiffer_Python_.py to FileDiffer.py
